Allow tests to be run via stack or cabal
Fixes #1584.

This commit introduces a new test option called '--liquid-runner' that
defaults to 'stack exec -- liquid' as before but allow users to override
the command which will eventually run 'liquid'.

Doing so allows Cabal users to pass something like 'cabal new-exec
liquid -- ' to the test executable, for example running this:

```
/path/to/test --liquid-runner 'cabal new-exec liquid -- '
```